Initial commit
Included are a few demo files of how to apply modulation of an LFO filter to sounds derived from triple oscillators.
In these examples, we use a harp sound from some LMMS plugin (forgot which), which can be reproduced with a triple oscillator.

The modulated fields are the LFO modulation strength knob, the LFO attack, and the LFO speed.
The more distorted of the two tracks is the one that includes LFO speed.